Added tests where multiple GC runs happen for a single successful summarization attempt due to retries (#18849)
#18632 added GC sweep ops. As called out in the Todo section of that PR,
if GC succeeds but a follow up step fails during summarization, the next
GC run may generate an sweep op that contains the same deleted nodes id.
This PR adds tests to validate these two scenarios:
1. GC succeeds but summary fails multiple times on summarization retries
until the final summary attempt which is successful. A sweep op is sent
during the first attempt which may get acked during the attempts. When
it does, not further sweep ops are sent.
2. GC succeeds but summary fails multiple times on summarization retries
until the final summary attempt which is successful. The sweep ops are
not processed (inbound queue is paused) so every GC attempt ends up
sending a sweep op with the same deleted node ids.

Also, update `summarizeOnDemand` to support running with retries in case
of failures.

AB#6547