Enable idCompressor for DOProviderContainerRuntimeFactory (#18961)
## Description

This turns on the id compressor for DOProviderContainerRuntimeFactory,
which allows using @fluidframework/tree with azure-client/odsp-client
after #18775 is merged.

A note on test changes: our jest tests have a 45 second beforeAll hook
which waits for an initial page load before running the tests. Since
id-compressor is dynamically imported, it's plausible webpack will serve
the main chunk (which includes a page load) before the whole contents of
the page are ready. If this happens, the tests can start running before
webpack finishes compiling/serving code capable of rendering the example
app. Luckily, the example tests were already set up with a mechanism for
when Fluid was actually ready to load. That check has been added to the
beforeAll hook to ensure tests are ready to run.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>