Code cleanup in container runtime around refresh from summary logic (#19002)
Cleaned up the following from container runtime. These were leftovers
from the logic to update the latest summary by downloading a snapshot.
The PR that changed the logic -
https://github.com/microsoft/FluidFramework/pull/17185.
Before the above PR, when `refreshLatestAck` summary option was true,
the latest snapshot was downloaded and the latest summary state was
updated from that. After the above PR, we never update the latest
summary state from a snapshot, instead we close the summarizer in this
scenario.

- `latestSnapshotVersionId` is no longer used in `submitSummary` and has
been removed.
- In `refreshLatestSummaryAck`, if a snapshot is fetched and is not the
latest, a data processing error is thrown. However, this is no longer
needed because the summarizer is closed in this scenario anyway. The
snapshot fetch is just an optimization and throwing an error is
redundant.