build(docs): Remove staging step when generating API docs (#19056)
The website build use to run a pre-processing step on the imported API
models before running docs generation. That logic was removed some time
back, but a build step remained that copied the doc models to a
"staging" directory before running the generation. This step was
unnecessary and is now removed.

Also updates some of the build scripts to add parallelism, and updates
script docs in the README.