build: Narrow workspace definitions (#19047)
Our workspace definitions are including projects that shouldn't be
included, especially when using temporary package.json files to control
the compilation of ESM/CJS in dual-emit packages.

This change updates the definitions to be narrower. For the client
workspace in particular, the results are more verbose because we have
packages at arbitrary depths in the tree. We should address this over
time by reorganizing the repo. For the server-related release groups the
change is much smaller because they're small and follow conventions.