Align detached data format with attached pending data (#18829)
Current serialize API returns a stringified ISummaryTree. In order to
consume it, rehydration receives that summary and converts it into an
ISnapshotWithBlobContents. This change can be summarized as moving that
summary -> snapshot conversion at serialization time. Such a conversion
was not exactly equal to current one in order to provide a format
similar to IPendingContainerState (ISnapshotTree and
ISerializableBlobContents instead of ISnapshotWithBlobContents).

Backward compatible.