build(docs): update render-api-documentation postfix logic and readVersionData (#19064)
Updated postfix logic so that render-api-documentation will properly
target api-extractor-temp-v1 instead of api-extractor-temp. Originally,
it was assumed that currentVersion would always represent the content
generated by api extractor. However, with the currentVersion set to v1,
additional conditions need to be checked.

Also updated readVersionData to retrieve from a new "downloadVersions"
field instead of "previousVersions". I realized that previousVersions
was not a good representation for the desired output of this script
since this script is intended to get versions which cannot be extracted
from pipeline outputs rather than all previous versions.