feat(docs): Update multi-version build to include API landing pages for each version (#19034)
Updates the website build to generate more user-friendly landing pages
for the root of each version of the API docs. Currently, only our "V1"
docs are live, but this PR adds a landing page for the "V2" docs,
currently marked as a draft. Once we are ready to enable the V2 docs, we
can remove the draft flag to enable it.

Before:

![image](https://github.com/microsoft/FluidFramework/assets/54606601/94a86f78-aaf5-424b-863a-3f0f1fb968ae)

After:

![image](https://github.com/microsoft/FluidFramework/assets/54606601/c85cacfb-dd61-4758-86b1-8cb226f77d6a)

V2 landing page:

![image](https://github.com/microsoft/FluidFramework/assets/54606601/dd1e28af-1e40-4577-9694-d1e5b59cb5e2)