Remove processing acks that are older than the one client loaded from (#19013)
Currently, when a summarizer client first loads, it will most likely
process the ack of the summary it loaded from. This is totally
unnecessary as this ack gets ignored. We can see this in telemetry today
as shown below where `refreshLatestSummary` logs are soon as the
Summarizer starts before it even attempts a summary:

![image](https://github.com/microsoft/FluidFramework/assets/45832642/04bf079d-ef32-47eb-919a-0952094747ef)

A summarizer client should only care about acks for summaries that it
generated itself. Summarizer client connects in "write" mode and it
waits for connected state before it creates `RunningSummarizer`. So, by
the time `RunningSummarizer` is created, all older acks will have been
processed.

This change removes the logic that processes the last ack seen when
`RunningSummarizer` starts. When it starts waiting for acks in a loop,
it starts waiting for acks from the summary next to the one it loaded
from. This will ensure that summarizer client never processes ack for
the summary it loaded from.

Note: It's possible in rare cases that while it was catching up, it saw
an ack that was newer than the one it loaded from and was from another
summarizer running in parallel. In this scenario, the summarizer closes
today during ack processesing flow because it never processes newer acks
from other summarizer clients. With this change, in this scenario, the
summarizer will instead close when uploading the summary because its
parent will be incorrect.