DDS: Tests illustrating which DDSes obscure handles in their op content (#18926)
We have moved the detection of handles for GC from the DDS layer
up to the ContainerRuntime. We can do this because we are soon
requiring DDSes to submit op payloads with handles unobscured.

This change adds a framework for testing each DDS that supports storing
handles, to illustrate whether the handle will be visible to the
ContainerRuntime layer.