Reduce access to IdCompressor (#19096)
## Description
The IdCompressor was being passed down rather than a RevisionTagCodec
where a RevisionTagCodec was sufficient. I've also added the
localSessionId to the codec so that we don't have to pass the compressor
or a sessionId through (avoids adding two parameters in a lot of cases).

Tests are still hooked up to an actual "global" compressor but a future
change will introduce a mock RevisionTagCodec that's a NoOp for tests
that shouldn't be testing compressor usage.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).