API(tree): make CheckoutEvents internal (#19118) (#19157)
Cherry-pick https://github.com/microsoft/FluidFramework/pull/19118.

This PR makes the `CheckoutEvents` interface internal and introduces
`TreeViewEvents` in its stead. `CheckoutEvents` is from the
`TreeCheckout` layer and was being used in the public API at the simple
tree layer. The new `TreeViewEvents` is solely meant to be used at the
simple tree layer. This reduce the coupling between the two layers,
allowing the event sets to be different (see below) and more precisely
documented.

This PR also introduces a change in the set of events being made public
(which is what motivated the above in the first place). The `revertible`
event is no longer advertised. This is because it the revertible system
is not yet complete, and whatever is currently present is not of beta
quality and not to spec. The sample apps have been updated to not rely
on it. This event API (and more) will be reintroduced at a later date
when ready and to spec.

## Breaking Changes

Since `CheckoutEvents` is no longer publicly exported, any references to
it will now fail to compile. `TreeViewEvents` should be used instead.
Any references to the following should be removed:
* `CheckoutEvents.revertible`
* `Revertible`
* `RevertibleKind`
* `RevertResult`
* `DiscardResult`