Fix Rebase with Inflight Ops on Merge Tree (#19080) - RELEASE 8.0 (#19200)
Porting https://github.com/microsoft/FluidFramework/pull/19080 to the
release branch.

Fix a bug in the rebase logic of merge tree where we weren't handling
out of order ops at the segment level. Specifically, the head of the
segment queue may not match the rebase op, as we allow there to be
outstanding ops before the rebase. The fix for this is fairly
straightforward, rather than dequeue we attempt to remove the segment
group, regardless of positions, but still validate it exists. ordering
will still be validated in the acknowledgement phase once the ops round
trip.

---------

Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>
Co-authored-by: Tony Murphy <anthonm@microsoft.com>