Handling of the MIP labels update from JoinSession and Fluid Signals from push and exposing them on container (#19142)
## Description

ADO Task Link:
https://dev.azure.com/fluidframework/internal/_workitems/edit/6640
Handling of the MIP labels update from JoinSession and Fluid Signals
from push and exposing them on container so that the apps can listen to
the updates and take appropriate action like update their UX etc.

The details about the server side changes etc. are included in the
design document.

1.) Add event " metadataUpdate " in IDocumentServiceEvents which are
emitted on IDocumentService.
**export interface IDocumentServiceEvents extends IEvent { 
/** 
* This event is used to communicate any metadata related to the
container. We might have received metadata from the service.
* Read more info on this event from here `IContainer.containerMetadata`.
*/ 
(event: "metadataUpdate", listener: (metadata: Record<string, string>)
=> void);
}** 

2.) DeltaConnection will extract the labels from the join session
response and will also extract it from the fluid signals in the driver
layer and whenever it is updated by either path, it will tell
IDocumentService to emit the ` metadataUpdate` event on the service
object.

3.) Container will listen to this event and then it will emit this info
on itself. For that we will add the same event in IContainerEvents.
**export interface IContainerEvents extends IEvent { 
/** 
* Emitted when the some of the properties related to the container are
initialized or updated.
* This emitted metadata will the props which are updated. If consumer
wants to read full set of metadata then they can read it off the
container from {@link IContainer.containerMetadata} prop.
*/ 
(event: " metadataUpdate ", listener: (metadata: Record<string, string>)
=> void);
}**  
Also expose containerMetadata: Record<string, string> on IContainer, so
that it can read at any time.

We are adding this prop on the container as user should be able to read
it on load also, besides getting any update through events when the
update happens. The emitted event will only have the props which are
updated while the contianerMetadata will have full set of props.

5.) Apps can listen to this event to get the latest policy labels. Apps
can also read this on load from the container object once the container
is connected. It will be exposed like this on the event or in
containerMetadata.

Container.containerMetadata = { sensitivityLabelsInfo: value(string)} 

Value will be jsonified string. So, app needs to parse it and it will
contain same structure as we received from the service which is:

{ labels: Object, timestamp: string } where labels is of structure:  

[ 
         { 
             "sensitivityLabelId": string, 
             "tenantId": string, 
             "assignmentMethod": string, 
         } 
 ]

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>