Remove deprecated SingleRT flow in odsp-driver (#18690)
This PR removes an old deprecated path of SingleRT feature in the
odsp-driver code. This code was marked as deprecated in
https://github.com/microsoft/FluidFramework/pull/11177 and was marked to
be removed with FF version 2.0.0-internal.1.1.0.

The code being removed corresponds to an earlier version of SingleRT
feature https://github.com/microsoft/FluidFramework/pull/7554 where
creation of odsp-file and creation of it's sharing link was done through
a single network call, by providing the enableShareLinkWithCreate
boolean flag to the HostStoragePolicy of odsp-driver.

The newer version of SingleRT feature
https://github.com/microsoft/FluidFramework/pull/11177 is still valid
and available to use in the odsp-driver code through
enableSingleRoundTripForShareLinkWithCreate boolean flag in
HostStoragePolicy. No new changes are done to it.

Motivation of this PR is to cleanup the code, before I make further
incremental changes in the same code path which I plan to implement in
the coming week