GC: Two-Stage Auto-recovery on Tombstone Load error (#19115)
Fixes AB#6234.

During the Grace Period, interpret usage of a Tombstoned object as an
added reference, so the unreferenced timestamp will be reset in the next
GC (See #18559 for info on Two-Stage Sweep). This will prevent Sweep
from deleting it, which is the primary goal.

Note: It's unlikely that this will _permanently_ repair an object that
is hitting Tombstone Loaded events. Two possibilities and outcomes:

1. There is truly no handle to this object, and some app logic is
wrongly referencing it. Until that app code is fixed up, this is liable
to repeat periodically (per the Tombstone Timeout). But it gives app
devs a chance to observe the bug while protecting end users from
permanent harm (assuming they hit this during the Grace Period)
2. There is a proper reference to the object, but the GC data was
miscalculated or corrupted. Unless the object with the reference changes
and has its GC data recomputed properly, this will also repeat
periodically. The plan is to make a subsequent PR that will also trigger
Full GC on the next GC run after the TombstoneLoaded op.