Added test config provider that provides options to set and clear config values (#19074)
Some of the GC and summary tests use a mock config provider and a
settings object to configure values. However, there is no consistent way
of clearing the settings value and as a result, the values may not be
cleared and new values may not be correctly set. There were a few
instances of this that were fixed recently.

This change adds a `ITestConfigProvider` interface which is an extension
of `IConfigProviderBase`. It adds methods `set` and `clear` which can be
used to set a configs value and clear all values respectively. Tests can
set values at the beginning of each test and reset all values when the
test ends. This way every test will start with a clean set of configs
and can add them as needed.

AB#6656