fix(matrix): Expand collab window to include refSeqs of in-flight ops (#19211)
## Description

Fixes a matrix resubmission bug, where zamboni might clean up
permutation segments which in-flight ops might need to refer to.
In general, using the minimum sequence number of the most recently
processed message in isolation to determine the collab window for a DDS
is not correct, since if that DDS has any in-flight ops and they get
resubmitted, they will need to be rebased from their refSeq to the last
processed deltaManager seq at the time of rebasing.

This bug seems to generally apply to DDSes which clean up collab-window
focused state. AB#6866 tracks fixing other DDSes / unifying the
mechanism used here.

While investigating this bug, I also found a bug related to the
double-reconnect scenario after adding matrix fuzz tests. This bug was
relatively easy to fix: the localOpMetadata used for a resubmitted op
needs to update its refSeq to be the current seq rather than reuse the
refSeq of the original submission time.

This resolves SharedMatrix-side issues for #18082.

AB#6356

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>
Co-authored-by: Connor Skees <39542938+connorskees@users.noreply.github.com>