GC: Remove unreferenced state tracking immediately upon detecting new reference (#19234)
The goal here is that when a reference is added (via
addedOutboundReference), we should immediately clear unreferenced state
tracking.

We can do this effectively and reliably from it because we detect added
references "inline"/immediately as the ContainerRuntime processes the op
stream. The benefit of removing unreferenced state tracking immediately
is that subsequent usages (changed/loaded) will not be wrongly logged
anymore.