refactor(map): Upgrade eslint config from "minimal" to "recommended" (#19269)
The "minimal" config is intended for prototype packages and repo-local
infra, it is not intended for use in production libraries.

This package was migrated to the "recommended" config in the past, but
at some point it was apparently downgraded back to "minimal", and a
number of new linter violations were introduced. This PR fixes those
violations and restores the config to the "recommended" level.