Converts build-api-markdown-documenter.yml pipeline to 1ES template (#19243)
## Description

This PR converts build-api-markdown-documenter.yml to the new 1ES
template required by Microsoft with the minimal required set of template
changes required. This new template enables security and auditing checks
by Microsoft. Read more on how the logic for converting to a 1ES
pipeline works on [the official 1ES
wiki](https://eng.ms/docs/cloud-ai-platform/devdiv/one-engineering-system-1es/1es-docs/1es-pipeline-templates/onboarding/overview)

A temporary 1ES folder was created for converted pipeline templates.
This enables other pipelines to remain unchanged and use the original
templates while new PR's are opened for individually converted other
pipelines to 1ES as required by Microsoft. **You can easily view what
the template changes are by looking at the diff in [the following
PR](https://github.com/microsoft/FluidFramework/pull/19244) which is a
fork of this branch and simply moves the template changes from the 1ES
folder to the original so there is a clear diff present on github**

- Once all pipelines are moved over, we can start moving the files out
of the 1ES folder and eventually remove the folder since it will no
longer have any files (This is after we convert the rest of the required
pipelines).