(odsp-client): Move file name and path closer to attachment flow (#18425)
This PR makes an attempt to bring `filePath` and `fileName` closer to
the attachment flow. Create odsp container requires file name and file
path which are required parameters while attaching a container.

The interface `IFluidContainer` function `attach` takes an optional
parameter of type `ContainerAttachProps`. This ensures a structured and
type-safe approach to attaching containers, offering flexibility for
clients to provide specific attach properties conforming to the defined
schema. This reduces the risk of passing unintended or incompatible
properties to the attach method.

Additionally, this change also allows passing custom file path and file
name.


[AB#6281](https://dev.azure.com/fluidframework/internal/_workitems/edit/6281)

[AB#6785](https://dev.azure.com/fluidframework/internal/_workitems/edit/6785)