ci(docs): Remove unneeded step in build:docs pipeline (#19291)
The npm script `build` (and by extension `ci:build) was previously
updated to perform the necessary steps to download and stage the
appropriate API model artifacts for API documentation generation. The
removed step here is a legacy one that used to serve this purpose, but
no longer does. The artifact being downloaded by this step was
completely ignored by the website build.