Unify eventing for interval endpoint and property changes  (#19206)
Creates a new event to emit whenever the endpoints, properties, or both
on an interval are changed. This is to better align with the changes
made in #18676 that combine the changing of endpoints and properties
into a single method. To avoid a breaking change, the new event is
emitted along with the existing changeInterval and propertyChanged
events.

AB#6556

---------

Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>