Only do back compat to loader v2 int 3 (#19304)
[AB#6724](https://dev.azure.com/fluidframework/internal/_workitems/edit/6724)

Previous change: https://github.com/microsoft/FluidFramework/pull/19167

We were still getting summary nacks from ODSP for tests that odsp loader
compat. Did not realize that it was the loader that was causing the
issue. This also makes it so for ODSP full back compat loader + driver
tests we run only from the current version to version
`2.0.0-internal.3.x.y`