Add explicit event when summarize fails all retries (#19209)
There currently isn't an explicit failure telemetry event when a
summarize fails all attempts, and as such dashboards are having to use
the `"Summarize_cancel"` event. However, there are certain expected
flows that will end up sending a `"Summarize_cancel"` telemetry event.
This event is seen as a "failure" when it could be a totally normal
flow.

In my mind, summarizer success/failure events should help answer 2
questions for dashboards and metrics:
1. What percentage of summaries are failing completely (no successful
retries)
2. How many retries are happening per success

In a real-world scenario, if a summarizer failed on 2 retries and
succeeded on the 3rd we would consider that a success. But our telemetry
is counting each individual retry as a failure, leading to improper
percentages of, for example, only succeeding 33% of the time.


[AB#6717](https://dev.azure.com/fluidframework/internal/_workitems/edit/6717)

**Note:** There currently exists the `"StoppingSummarizer"` event, but
this happens every time the summarizer stops for normal reasons.
Dashboards and metrics would need to be built off the [`"reason"`
property](https://github.com/kian-thompson/FluidFramework/blob/main/packages/runtime/container-runtime/src/summary/summarizerTypes.ts#L379),
which is subject to change and probably not the best way forward.