Reapply fix(matrix): Expand collab window to include refSeqs of in-flight ops (#19302)
## Description

Reapplies the fix in #19211 with some changes to how strict we are with
the queue of in-flight ops. Notably, it appears that
`this.runtime.deltaManager.lastSequenceNumber` at op submit time
sometimes trails behind the reference sequence number that actually gets
submitted for that op, as evidenced by some e2e tests. This had
something to do with client leave/join messages. Rather than assert
equality on round-trip, I've converted the mismatch assert to verify
`<=`, since that's all that's required for correctness of the scheme.

The applyStashedOp codepath also happened to work before but in an
unexpected way: `reSubmit` was calling `.shift()` on an empty queue.
This PR adds enforcement to the reSubmit codepath and a corresponding
unit test for that flow.

The changes here are split naturally into commits for reviewers familiar
with that PR.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>