Add cluster draining interface (#18975)
This PR includes changes below:
1. Create cluster draining checker interface to get cluster draining
status
2. Add draining status check in Alfred get session API to reject the
request if cluster is in draining. It is to prevent clients keep sending
request to cluster under draining and mess up the session flags
3. Add draining status check in connect document call to block new
connect_document request when cluster is draining
4. Add draining status check in deli lambda close flow to add skip
session stickiness flag. This is to ignore stickiness and allow session
to be moved to a new cluster
5. Change get session API to move session when sticky if the skip
stickiness flag is true

---------

Co-authored-by: Yunpeng Hou <yunho-microsoft@Yunho-MacBookPro2019.local>
Co-authored-by: Yunpeng Hou <yunho-microsoft@YunhoMakPro2019.guest.corp.microsoft.com>