build(docs): Enable dynamic wildcard redirects on FF.com (#19140)
The PR enables wildcard redirects in FF.com. When a user goes to a link
that starts with `/docs/apis/*` it should redirect them to
`/docs/api/<current-version>/*`. `/docs/api/current/*` and
`/docs/apis/lts/*` are also included in this PR

## Instructions to reproduce the redirects;
1. `cd docs/ && pnpm i && pnpm run build && npm i -g
@azure/static-web-apps-cli`
1. `swa start public --api-location api`
1. Modify the url from `/docs/api/<current-version>/` to `/docs/apis/`
and see it go back to `/docs/api/<current-version>/`