tree: Pairwise composition of changesets (#19311)
## Description

This PR changes the signature of `FieldChangeRebaser.compose` to take
two changesets to compose instead of an array of changesets. Before this
PR, field kinds could call `composeChild` on a given `NodeChangeset`
multiple times during a single compose pass. This violated assumptions
made in `ModularChangeFamily`'s scheme for amending fields when new
relevant cross-field information is found. Fixing this issue will enable
a future PR to remove `amendCompose` and switch to a simpler scheme
where field composition is recomputed instead of amended.

---------

Co-authored-by: yann-achard-MS <97201204+yann-achard-MS@users.noreply.github.com>