Enable all getAttachSummary paths in Legacy Shared Tree (#19365)
[AB#7049](https://dev.azure.com/fluidframework/internal/_workitems/edit/7049)

Original change causing the regression:
https://github.com/microsoft/FluidFramework/pull/18088

A customer was seeing `0x62e` errors appear in telemetry after
`2.0.0-internal.7.4.3` was consumed. After an investigation, a
regression was detected. It was determined that the original fix did not
properly check if an attach summary was being generated in all cases,
and would fail in some cases when an attach summary was generated.

When a datastore was attached, customers could not create new legacy
shared tree DDSes with local edits as assert `0x62e` would be hit as the
check would check to see if the datastore was attached, and not if the
DDS was "attached".

When a datastore was detached, while attaching the datastore, the assert
`0x62e` would be hit as the datastore runtime would be put into an
`attaching` state thus the legacy SharedTree DDS would consider itself
`attached` when the attach summary was generated.

Instead of putting the local edit logic when generating both attach
summaries and regular summaries, we put the local edit logic only when
generating the attach summary. This way, we already know that the legacy
SharedTree DDS is being attached. When generating a regular summary, the
assert `0x62f` should fire off if there are any local edits to a summary
when it's being generated.