build(docs): Setting skip_api_build to true in docs pipeline (#19417)
Pipeline was failing on deployment of FF.com after adding the redirects
code. Since the azure function is written in js, it does not need a
build step like it would if its written in TypeScript, and so node isnt
needed. Setting skip_api_build to true resolved this bug