Summarizer ITelemetryContext: Deprecate "output" functions get/serialize (#19409)
`ITelemetryContext` is passed all over the place in the runtime to allow
layers to jot down info for instrumenting Summarization. It should be a
"write-only" interface like the loggers (they only have `send`, not
anything about reading what's been logged).

Like all our internal instrumentation APIs, we want to use the `Ext`
logger types which are richer than the public interfaces. Unfortunately,
`ITelemetryContext` has to be public for now, but at least if we make it
"write-only" then the telemetry props will tolerate any change to the
`Ext` props type that expands what's allowed (all old callsites will
still work). But if we expose getters that _return_ these props, and we
want to expand that type, then it's a non-back-compat breaking change
and has to wait for the next major release!