Deprecate PureDataObjectFactory.createRootInstance (#19428)
[AB#7082](https://dev.azure.com/fluidframework/internal/_workitems/edit/7082)

## Description
Deprecate `PureDataObjectFactory.createRootInstance` as a follow up to
creating `PureDataObject.createInstanceWithDataStore`. Create root
instance was updated to use aliasing, but it doesn't give customers the
freedom to control what happens when an aliasing conflict occurs.

For the current `createRootInstance` assumes that when an aliasing
conflict occurs, it occurs because the same `DataObject` type was
already created with the same alias. If a customer wants a behavior that
returns a different type of object, that is not possible with
`createRootInstance`. It however is possible with
`createInstanceWithDataStore`.

Originally the `createRootInstance` used older deprecated apis and would
not properly handle this, and thus we created the aliasing feature.

## Considerations
We can also deprecate `createPeerInstance` and `createChildInstance`,
however all this does is fix the package path recorded in the
`IFluidDataStoreContext`. There isn't really any "unique" benefits for
the customer other than from a debugging standpoint. This same effect
can now be achieved with the `createInstanceWithDatastore` api.

`createChildInstance` is used in our examples, so in order to deprecate
them we would need to remove `createChildInstance` from our examples.