Initial work for summarizer fuzz testing (#19404)
We want to enable fuzz testing for summarizer, and this PR starts that
work.

Summarizer fuzz testing should test that we eventually recover and send
a summary successfully.
For DDS, we test for eventual consistency. For summarizer, we could test
for eventual recovery.

After performing operations (i.e. disconnects, summaryNacks, ops from
other clients, etc.) we should:
- start a fresh summarizer
- attempt a summary
If the system doesn't recover properly, then we have a bug to fix


[AB#6919](https://dev.azure.com/fluidframework/internal/_workitems/edit/6919)