Refresh immediately after summary ack (#18968)
[AB#6490](https://dev.azure.com/fluidframework/internal/_workitems/edit/6490)

## Background
Attempting
[AB#3400](https://dev.azure.com/fluidframework/internal/_workitems/edit/3400)
to add validation to the summarizer before summaries occurred to verify
that the summary the running summarizer is tracking is the same as the
summarizer node system, we discovered that the validation was failing.
The validation caught locally that a single summarizer would start a
second summary before processing the ack of the first summary it
created.

## Solution
This PR attempts to address that issue. Instead of waiting for the
summary lock to be released, the ack is immediately refreshed. This may
result in multiple refreshes. The `ContainerRuntime`'s refresh logic is
resilient and should be able to handle refreshing multiple of the same
acks.

## Further Work
The `RunningSummarizer` really only expects its own acks. Any foreign
acks should close/restart the summarizer. The code in the
`RunningSummarizer` should reflect that reality.