GC: Detect references added via attach ops (#19275)
Fixes AB#5757

With this PR we fill a gap in detecting references as they come - in the
case where an object is re-referenced in a new object (the first
reference appears in that attach op). _Background: This is important in
the case where that reference is quickly removed (before the next GC
run). Otherwise we would not update the unreferenced timestamp and GC
calculations would be incorrect._

Just as we include the container's GC state in each summary, we need to
include an object's initial GC Data (list of references) in its attach
summary. This data can be parsed out of the attach message on process
and GC can be notified.

Existing coverage:

- GC reference detection for DataStore attach:
`gcUnreferencedTimestamp.spec.ts`
- Overall attach flow (op generation and processing), including compat
between runtimes: `attachLifecycle.spec.ts`

Additional coverage:

- `gcUnreferencedTimestamp.spec.ts`:
  - Use a DDS in the chain
  - Use multiple DDSes in the chain
- `gcInactiveNodes.spec.ts` - Revival happens on local client that sent
the attach op
  - DataStore attach
  - DDS attach